Add _makeRemoveFilter to removeTranslation
git-svn-id: http://svn.php.net/repository/pear/packages/LiveUser_Admin/trunk@181543 c90b9560-bf6c-de11-be94-00142212c4b1

+++ Perm/Simple.php
@@ -559,6 +559,11 @@ class LiveUser_Admin_Perm_Simple
      */
     function removeTranslation($filters)
     {
+        $filters = $this->_makeRemoveFilter($filters, 'section_id', 'getTranslation');
+        if (!$filters) {
+            return $filters;
+        }
+
         $result = $this->_storage->delete('translations', $filters);
         // notify observer
         return $result;
